+++ package.json
@@ -46,7 +46,7 @@
     "prepare": "cd app && npm install && cd .. && npm run build",
     "test:integration": "jest --testRegex '.*integration-test.js'",
     "test:manual": "npm run build && ./.github/manual-test",
-    "test:spectron": "jest  --testRegex '.*spectron-test.js'",
+    "test:playwright": "jest  --testRegex '.*playwright-test.js'",
     "test:unit": "jest",
     "test:watch": "echo 'Remember to run npm run build:watch for the test watcher to work!' && jest --watchAll --collectCoverage=false",
     "test:withlog": "LOGLEVEL=trace npm run test",
@@ -81,9 +81,9 @@
     "eslint-config-prettier": "^8.1.0",
     "eslint-plugin-prettier": "^3.3.1",
     "jest": "^27.0.3",
+    "playwright": "^1.12.1",
     "prettier": "^2.2.1",
     "rimraf": "^3.0.2",
-    "spectron": "^14.0.0",
     "ts-loader": "^9.2.1",
     "typescript": "^4.2.3",
     "webpack": "^5.24.4",
